Change note about where to edit client_max_body_size
When I did as the doc said and placed an updated client_max_body_size directive in the "server" override I got the following error, '"client_max_body_size" directive is duplicate' when restarting nginx. Placing it in the "location" override and restarting did not cause this error during restart and correctly applied the updated configuration after the restart.